Added basic camera movement
I added basic camera movement in this push so it is easier to explore larger maps which supports creating a larger game world.

How input is handled in the current form is not final, eventually this functionality is intended to be contained inside the input handler class alongside having a custom Event System that can be be published by a component and then will send out those input events that can be subscribed to by any interested components.