Fixed tilemap rendering exception ( See description for details)
Fixed tilemap rendering exception that occurs when a tilemap is of a non-square configuration. To get the index of the tile to render in the array it would multiply the x value by the width then add the y. This caused it to overflow when in a non-square configuration.

It now instead multiplies the y value by the width and adds the x value.