Tilemap refactor (Separated from engine code)
Separated the actual creation of the tilemap from the engine code as it'll need to be easily accessible for any game using it. Mainly because any game using it will need to be easily able to modify a single tile and will be accessing it alot so it makes sense that the ownership be with the game code itself.

I took the opportunity to make it so that the base scene class instead can store a list of Tilemap pointers to render to give the engine the ability to render multiple tilemaps (1 for environment, 1 for units etc...).

The list takes in raw pointers as this list should not own the Tilemaps and therefore it doesn't need to worry about ensuring their memory is freed upon no longer being used.